+++ modules/actor-sheet.js
@@ -42,6 +42,7 @@ export class witchcraftActorSheet extends ActorSheet {
         const power = [];
         const quality = [];
         const skill = [];
+        const specialty = [];
         const drawback = [];
 
         // Iterate through items and assign to containers
@@ -68,6 +69,10 @@ export class witchcraftActorSheet extends ActorSheet {
                     skill.push(i)
                     break
 
+                case "specialty":
+                    specialty.push(i)
+                    break
+                    
                 case "drawback":
                     drawback.push(i)
                     break
@@ -75,7 +80,7 @@ export class witchcraftActorSheet extends ActorSheet {
         }
 
         // Alphabetically sort all items
-        const itemCats = [item, equippedItem, weapon, power, quality, skill, drawback]
+        const itemCats = [item, equippedItem, weapon, power, quality, skill, specialty, drawback]
         for (let category of itemCats) {
             if (category.length > 1) {
                 category.sort((a, b) => {
@@ -94,6 +99,7 @@ export class witchcraftActorSheet extends ActorSheet {
         actorData.power = power
         actorData.quality = quality
         actorData.skill = skill
+		actorData.specialty = specialty
         actorData.drawback = drawback
     }
 
@@ -189,11 +195,17 @@ export class witchcraftActorSheet extends ActorSheet {
         let attributeLabel = element.dataset.attributeName
         let actorData = this.actor.system
 
-        // Create options for Qualities/Drawbacks/Skills/Powers
+        // Create options for Qualities/Drawbacks/Skills/Specialties/Powers
         let skillOptions = []
         for (let skill of this.actor.items.filter(item => item.type === 'skill')) {
             let option = `<option value="${skill.id}">${skill.name} ${skill.system.level}</option>`
             skillOptions.push(option)
+        }
+        
+        let specialtyOptions = []
+        for (let specialty of this.actor.items.filter(item => item.type === 'specialty')) {
+            let option = `<option value="${specialty.id}">${specialty.name}</option>`
+            specialtyOptions.push(option)
         }
 
         let qualityOptions = []
@@ -269,6 +281,15 @@ export class witchcraftActorSheet extends ActorSheet {
                                             </select>
                                         </td>
                                     </tr>
+									<tr>
+                                        <td class="table-bold-text">Specialties</td>
+                                        <td>
+                                            <select id="specialtySelect" name="specialties">
+                                                <option value="None">None</option>
+                                                ${specialtyOptions.join('')}
+                                            </select>
+                                        </td>
+                                    </tr>
                                     <tr>
                                         <td class="table-bold-text">Qualities</td>
                                         <td>
@@ -311,6 +332,7 @@ export class witchcraftActorSheet extends ActorSheet {
                         let attributeTestSelect = html[0].querySelector('#attributeTestSelect').value
                         let userInputModifier = Number(html[0].querySelector('#inputModifier').value)
                         let selectedSkill = this.actor.getEmbeddedDocument("Item", html[0].querySelector('#skillSelect').value)
+                        let selectedSpecialty = this.actor.getEmbeddedDocument("Item", html[0].querySelector('#specialtySelect').value)
                         let selectedQuality = this.actor.getEmbeddedDocument("Item", html[0].querySelector('#qualitySelect').value)
                         let selectedDrawback = this.actor.getEmbeddedDocument("Item", html[0].querySelector('#drawbackSelect').value)
                         let selectedPower = this.actor.getEmbeddedDocument("Item", html[0].querySelector('#powerSelect').value)
@@ -318,13 +340,14 @@ export class witchcraftActorSheet extends ActorSheet {
                         // Set values for options
                         let attributeValue = attributeTestSelect === 'Simple' ? actorData.primaryAttributes[attributeLabel.toLowerCase()].value * 2 : actorData.primaryAttributes[attributeLabel.toLowerCase()].value
                         let skillValue = selectedSkill != undefined ? selectedSkill.system.level : 0
-                        let qualityValue = selectedQuality != undefined ? selectedQuality.system.bonus : 0
+                        let specialtyValue = selectedSpecialty != undefined ? 2 : 0
+						let qualityValue = selectedQuality != undefined ? selectedQuality.system.bonus : 0
                         let drawbackValue = selectedDrawback != undefined ? selectedDrawback.system.bonus : 0
                         let powerValue = selectedPower != undefined ? selectedPower.system.level : 0
                         let statusPenalties = actorData.secondaryAttributes.endurance_points.loss_penalty + actorData.secondaryAttributes.essence.loss_penalty
 
                         // Calculate total modifier to roll
-                        let rollMod = (attributeValue + skillValue + qualityValue + powerValue + userInputModifier) - Math.abs(drawbackValue) + statusPenalties
+                        let rollMod = (attributeValue + skillValue + specialtyValue + qualityValue + powerValue + userInputModifier) - Math.abs(drawbackValue) + statusPenalties
 
                         // Roll Dice
                         let roll = await new Roll('1d10').evaluate()
@@ -342,6 +365,9 @@ export class witchcraftActorSheet extends ActorSheet {
                         if (selectedSkill != undefined) { 
                             const skillLevel = selectedSkill.system.level;
                             tags.push(`<span class="${skillLevel >= 0 ? 'bonusColorClass' : 'penaltyColorClass'}">${selectedSkill.name} ${skillLevel >= 0 ? '+' : ''}${skillLevel}</span>`) 
+                        }
+                        if (selectedSpecialty != undefined) { 
+                            tags.push(`<span class="${'bonusColorClass'}">${selectedSpecialty.name} ${'+2'}</span>`) 
                         }
                         if (selectedQuality != undefined) { 
                             const qualityBonus = selectedQuality.system.bonus;